exercise printable library

+++ src/main/scala/sandbox/typeclass/Json.scala
@@ -4,7 +4,7 @@ package sandbox.typeclass
 sealed trait Json
 final case class JsObject(get: Map[String, Json]) extends Json
 final case class JsString(get:String) extends Json
-final case class JsNumber(get:Double) extends Json
+final case class JsNumber(get:Int) extends Json
 case object JsNull extends Json
 
 // serialize to Json
